+++ Packages/BrandonUtils/Standalone/Optional/OptionalException.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+using JetBrains.Annotations;
+
+namespace BrandonUtils.Standalone.Optional {
+    internal static class OptionalException {
+        public static InvalidOperationException IsEmptyException<T>([NotNull] IOptional<T> self) {
+            return new InvalidOperationException($"Unable to retrieve the {nameof(self.Value)} from the {self.GetType().Name} because it is empty!");
+        }
+    }
+}