+++ Packages/BrandonUtils/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.14.1] - 9/2/2021
+
+### Changed
+
+- Added `IReadOnlyCollection<T>` to `IOptional<T>`
+- Added `[CanBeNull]` to `IOptional<T>.Value`
+
+### Fixed
+
+- Equality comparisons for `null` `IOptional<T>`s and `IOptional<T>`s _containing_ `null`
+
 ## [2.13.4] - 8/31/2021
 
 ### Changed